Implement removeFolder method to Storage

+++ src/main/stores/Storage.ts
@@ -1,5 +1,6 @@
 import { observable, action } from 'mobx'
 import { Note, Folder } from '../lib/db/dataTypes'
+import { FOLDER_ID_PREFIX } from '../../lib/consts'
 
 export default class Storage {
   @observable folderMap: Map<string, Folder> = new Map()
@@ -13,8 +14,26 @@ export default class Storage {
   }
 
   @action
-  removeFolder(id: string) {
-    this.folderMap.delete(id)
+  removeFolder(folderPath: string) {
+    const folderId = `${FOLDER_ID_PREFIX}${folderPath}`
+    const allFolderIds = [...this.folderMap.keys()]
+    const deletedFolderKeys = [] as string[]
+    allFolderIds.forEach(id => {
+      if (folderId === id || id.startsWith(`${folderId}/`)) {
+        deletedFolderKeys.push(id)
+        this.folderMap.delete(id)
+      }
+    })
+
+    const noteEntries = [...this.noteMap.entries()]
+    noteEntries.forEach(([noteId, note]) => {
+      if (
+        note.folder === folderPath ||
+        note.folder.startsWith(`${folderPath}/`)
+      ) {
+        this.noteMap.delete(noteId)
+      }
+    })
   }
 
   @action
